ipc_dashboard/requirements.txt
streamlit==1.43.0
plotly==5.24.1
pandas==2.2.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
